refactor: move endpoints to separate file
